+++ server/_custom/dbi.js
@@ -1,0 +1,64 @@
+/** ****************************************************************************
+ *
+ * Overwrites the methods of './server/dbi/dbi.js'.
+ *
+ * dbi.js is just a literal object that contains a set of functions.
+ * It can't be instantiated.
+ *
+ * Private Functions:
+ *  . none
+ *
+ *
+ * Public Methods:
+ *  . zzz                         ...,
+ *
+ *
+ *
+ * @namespace    -
+ * @dependencies none
+ * @exports      -
+ * @author       -
+ * @since        0.0.0
+ * @version      -
+ * ************************************************************************** */
+/* eslint one-var: 0, semi-style: 0, no-underscore-dangle: 0 */
+
+
+// -- Vendor Modules
+
+
+// -- Local Modules
+
+
+// -- Local Constants
+
+
+// -- Local Variables
+
+
+// -- Private Functions --------------------------------------------------------
+// none,
+
+
+// -- Public Methods -----------------------------------------------------------
+
+const methods = {
+
+  /**
+   * Does ...
+   * (must ...)
+   *
+   * @method ()
+   * @public
+   * @param {}              -,
+   * @returns {}            -,
+   * @since 0.0.0
+   */
+  zzz() {
+    //
+  },
+};
+
+
+// -- Export
+module.exports = methods;